requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# LangChain Core
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-google-genai>=0.0.5

# Vector Store
chromadb>=0.4.0

# Document Processing
pypdf>=4.0.0

# Environment Variables
python-dotenv>=1.0.0

# Additional dependencies
pydantic>=2.0.0
openai>=1.10.0
